+++ ThirdPartyAdapters/line/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## LINE Android Mediation Adapter Changelog
 
+#### Next Version
+- Removed uses of deprecated FiveAdViewEventListener to use 
+FiveAdCustomLayoutEventListener, FiveAdVideoRewardEventListener, 
+FiveAdInterstitialEventListener and FiveAdNativeEventListener.
+- Verified compatibility with FiveAd SDK version 2.7.20240112.
+- Updated the minimum required Google Mobile Ads SDK version to 22.6.0.
+
 #### Version 2.6.20230607.1
 - Added `LineExtras` class for publishers to define default sound settings for
 banner, interstitial and rewarded ad formats.